Updated based on code review A little more cleanup
